feat: #186 performance-test 프로필 및 관련 객체 구현
