+++ src/main/java/site/dogether/s3/url_generator/PresignedUrlGenerator.java
@@ -1,0 +1,6 @@
+package site.dogether.s3.url_generator;
+
+public interface PresignedUrlGenerator {
+
+    String generate(Long dailyTodoId, String uploadFileType);
+}
