+++ src/main/resources/application-performance-test.yml
@@ -1,0 +1,14 @@
+server.port: ${APPLICATION_PORT}
+
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: ${DB_URL}
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
+
+  jpa.hibernate.ddl-auto: create
+
+logging:
+  config: classpath:logback/logback-performance-test.xml
+  level.site.dogether: trace
